authority_ai_service_backend/face-detection-pipeline/requirements.txt
# Core ML & DeepFace
deepface==0.0.92
tf-keras
opencv-python
opencv-contrib-python
numpy>=1.24.3
Pillow>=10.1.0

# Face detection models
retinaface-pytorch==0.0.3
facenet-pytorch==2.5.3
mediapipe>=0.10.0

# API framework
fastapi>=0.100.0
uvicorn[standard]
python-multipart
requests

# Utilities
python-dotenv
pydantic>=2.0.0
pydantic-settings

# LangChain & LangGraph
langchain-core
langchain-community
langgraph
langchain-ollama

# Google APIs (for Gmail)
google-auth-oauthlib
google-api-python-client
google-auth-httplib2

# Testing
pytest
pytest-asyncio
httpx
